+++ cmd/frontend/graphqlbackend/user.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"strings"
 	"time"
 
 	graphql "github.com/graph-gophers/graphql-go"
@@ -17,12 +18,44 @@ import (
 	"github.com/sourcegraph/sourcegraph/pkg/errcode"
 )
 
-func (r *schemaResolver) User(ctx context.Context, args struct{ Username string }) (*UserResolver, error) {
-	user, err := db.Users.GetByUsername(ctx, args.Username)
+func (r *schemaResolver) User(ctx context.Context, args struct {
+	Username        *string
+	UsernameOrEmail *string
+}) (*UserResolver, error) {
+	var query string
+	if args.Username != nil {
+		// Deprecated query by "UsernameOrEmail"
+		query = *args.Username
+	} else if args.UsernameOrEmail != nil {
+		query = *args.UsernameOrEmail
+	}
+	// Check if querying for email address.
+	if strings.Contains(query, "@") {
+		user, err := db.Users.GetByVerifiedEmail(ctx, query)
+		if err != nil {
+			return nil, safeErrMsg(ctx, err)
+		}
+		// 🚨 SECURITY: Only the user and admins are allowed to access the email address.
+		if err := backend.CheckSiteAdminOrSameUser(ctx, user.ID); err != nil {
+			return nil, safeErrMsg(ctx, err)
+		}
+		return &UserResolver{user: user}, nil
+	}
+
+	user, err := db.Users.GetByUsername(ctx, query)
 	if err != nil {
 		return nil, err
 	}
 	return &UserResolver{user: user}, nil
+}
+
+// 🚨 SECURITY: safeErrMsg is an error message that can be shown to any user.
+// If the user is a site admin, the passed in error is returned.
+func safeErrMsg(ctx context.Context, err error) error {
+	if isSiteAdminErr := backend.CheckCurrentUserIsSiteAdmin(ctx); isSiteAdminErr != nil {
+		return isSiteAdminErr
+	}
+	return err
 }
 
 // UserResolver implements the GraphQL User type.